use boost mersenne twister rng, add boost::random dependency

+++ src/game_events/action_wml.cpp
@@ -77,6 +77,7 @@
 #include <boost/scoped_ptr.hpp>
 #include <boost/assign/list_of.hpp>
 
+
 static lg::log_domain log_engine("engine");
 #define DBG_NG LOG_STREAM(debug, log_engine)
 #define LOG_NG LOG_STREAM(info, log_engine)
@@ -2012,13 +2013,13 @@ WML_HANDLER_FUNCTION(set_variable, /*event_info*/, cfg)
 					<< 0x3fffffff
 					<< ".\n";
 			}
-			long choice = random_new::generator->next_random();// gameinfo->rng().get_next_random();
+			uint32_t choice = random_new::generator->next_random();// gameinfo->rng().get_next_random();
 			if(num_choices >= 32768) {
 				choice <<= 15;
 				choice += random_new::generator->next_random();//gameinfo->rng().get_next_random();
 			}
 			choice %= num_choices;
-			long tmp = 0;
+			uint32_t tmp = 0;
 			for(size_t i = 0; i < ranges.size(); ++i) {
 				tmp += (ranges[i].second - ranges[i].first) + 1;
 				if (tmp > choice) {
